Downloads/Developer-Research-AI-Agent-updated-__source_repo__/Developer-Research-AI-Agent-updated-__source_repo__/prompts/pyproject.toml
[project]
name = "ai-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search==6.3.2",
    "fastembed>=0.7.4",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.9",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.17.0",
    "streamlit>=1.54.0",
    "tavily-python>=0.7.21",
]
